fixes overflow bugs in itp pipelines

+++ etl/cite/itp_cite_empresas_tipo.py
@@ -93,7 +93,7 @@ class CiteEmpresasPipeline(EasyPipeline):
         dtypes = {
             'cite_id':               'UInt8',
             'empresa_id':            'UInt8',
-            'time':                  'UInt16',
+            'time':                  'UInt32',
             'empresas':              'Float32',
          }
